refactor: change the context usage structure
