--- src/context/InputContext.tsx
@@ -1,13 +1,0 @@
-import { createContext, Dispatch, SetStateAction } from 'react';
-
-interface InputContextProps {
-  inputValue: string;
-  setInputValue: Dispatch<SetStateAction<string>>;
-}
-
-const InputContext = createContext<InputContextProps>({
-  inputValue: '',
-  setInputValue: () => {},
-});
-
-export default InputContext;
